% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{init}
\alias{init}
\title{Various parameter initialization methods for use with the rpl_optim function}
\usage{
init(mydata, type = "unstructured", pi = NULL, block_indices = NULL)
}
\arguments{
\item{mydata}{Data matrix}

\item{type}{Type of intialization to perform, according to the desired
structure of correlation matrix:
\code{unstructured} ((d^2 - d)/2 correlation parameters estimated using
the empirical Pearson correlation matrix),
\code{exchangeable} (1 correlation parameter estimated by averaging off-diagonal
of the empirical Pearson correlation matrix),
\code{one-factor} (d theta parameters, where rho_{ij} = theta_i * theta_j,
estimated using Nelson-Meader optimization of fit compared to the estimated
Pearson correlation matrix),
\code{block_exchangeable} (for b blocks, (b^2 - b)/2 correlation parameters
estimated by averaging empirical Pearson correlation matrix within each block).
In all cases, marginal parameters are initalized using the plug-in method
with empirical marginal means.
Several alternative initializations are also proposed for unstructured
correlation matrices: PPearson correlation (\code{unstructured_A},
equivalent to \code{unstructured}), pairwise
optimization with all data (\code{unstructured_B}), pairwise
optimization with random subsample of data (\code{unstructured_C}),
random initialization for the copula parameters.
(\code{unstructured_D})}

\item{pi}{Bernoulli sampling parameter (0 < pi <= 1)}

\item{block_indices}{For \code{type="block_exchangeable"}, list
containing the indices of variables belonging to each of the b blocks.}
}
\value{
Vector of parameter values to be used to initialize the
\code{rpl_optim} function.
}
\description{
Various parameter initialization methods for use with the rpl_optim function
}
