% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{rpl-package}
\alias{rpl-package}
\title{Randomized pairwise likelihood method for complex statistical inferences}
\description{
Randomized pairwise likelihood approach to enable computationally
efficient statistical inference, including the construction of
confidence intervals, in cases where the full likelihood is too
complex to be used (e.g., multivariate count data).
}
\details{
\tabular{ll}{ Package: \tab rpl\cr Type: \tab Package\cr Version:
\tab 0.0.6\cr Date: \tab 2023-03-29\cr License: \tab GPL (>=3.3.1)\cr LazyLoad:
\tab no\cr }
}
\references{
Mazo, G., Karlis, D., Rau, A. (2021) A randomized pairwise likelihood
method for complex statistical inferences. In revision. hal-03126621.
}
\author{
Gildas Mazo, Dimitris Karlis, Andrea Rau

Maintainer: Andrea Rau <\url{andrea.rau@inrae.fr}>
}
\keyword{multivariate}
