% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rpl_loglike}
\alias{rpl_loglike}
\title{Fisher transformation applied within function}
\usage{
rpl_loglike(
  par,
  mydata,
  pi = 1,
  offsets = NULL,
  block_indices = NULL,
  seed = NULL,
  count_pairs = FALSE,
  parallel = FALSE,
  BPPARAM = ifelse(parallel, bpparam(), NULL)
)
}
\arguments{
\item{par}{Vector providing parameter values (means followed by correlations)}

\item{mydata}{Data matrix}

\item{pi}{Bernoulli sampling parameter (0 < pi <= 1)}

\item{offsets}{If needed, a matrix of offsets (optional)}

\item{block_indices}{List of vectors providing indices for each block
for blockwise exchangeable correlation matrices}

\item{seed}{Seed to use for random number generation}

\item{count_pairs}{If \code{TRUE}, return the number of observations retained for
each pair of variables used in the randomized pairwise likelihood}

\item{parallel}{If \code{TRUE}, some computations are parallelized with the
BiocParallel package}

\item{BPPARAM}{If \code{parallel=TRUE}, the registry of back-ends for use
in parallelized calculations}
}
\value{
Negative log-likelihood for user-provided parameter vector,
calculated with the randomized pairwise likelihood approach.
If \code{count_pairs = TRUE}, counts of the number of observations used
for each pair are also returned.
}
\description{
par is a vector of mean parameters followed by Fisher-transformed
correlation parameters
}
\references{
Mazo, G., Karlis, D., Rau, A. (2021) A randomized pairwise likelihood
method for complex statistical inferences. In revision. hal-03126621.
}
