% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_mvt_poisson}
\alias{simulate_mvt_poisson}
\title{Simulate multivariate Poisson data with a given correlation
matrix structure using Gaussian copulas}
\usage{
simulate_mvt_poisson(
  n = 1000,
  d = 4,
  param,
  offsets = NULL,
  type = "unstructured",
  block_indices = NULL
)
}
\arguments{
\item{n}{Number of observations to simulate}

\item{d}{Dimension of variables to simulate}

\item{param}{Named list (\code{"margins"}, \code{"disp"})
of parameters to be used for simulation. Parameters
should first include marginal parameters, then vectorized correlation parameters
in lexicographical order. The size of the \code{disp} vector depends
on the \code{type} of correlation matrix structure to be simulated.}

\item{offsets}{If needed, a matrix of offsets (optional)}

\item{type}{Type of correlation matrix structure to be simulated:
\code{unstructured} ((d^2 - d)/2 correlation parameters),
\code{exchangeable} (1 correlation parameter),
\code{one-factor} (d theta parameters, where rho_{ij} = theta_i * theta_j),
\code{block_exchangeable} (for b blocks, (b^2 - b)/2 correlation parameters).}

\item{block_indices}{For \code{type="block_exchangeable"}, list
containing the indices of variables belonging to each of the b blocks.}
}
\value{
\item{data }{Matrix of simulated data}
\item{param }{True parameter values used to simulate data}
}
\description{
Simulate multivariate Poisson data with a given correlation
matrix structure using Gaussian copulas
}
